Solved Errors
Change into SRC package
1) Create a package model And Pass Account class into it

2) and delete the Weithdraw Class

3) Next commit is for javaBeans change into the model class convert into java bean class
