+++ Ass_4/src/AccountOpDemo/AccountOverdrawDemo.java
@@ -1,27 +1,33 @@
-public class AccountOverdrawSafeDemo implements Runnable{
-    private Account account;
+package AccountOpDemo;
 
-    public AccountOverdrawSafeDemo(Account account) {
+import Model.Account;
+
+public class AccountOverdrawDemo implements Runnable{
+    private Account account;
+    //AccountOpDemo.Withdraw w=new AccountOpDemo.Withdraw();
+
+    public AccountOverdrawDemo(Account account) {
         this.account = account;
     }
 
     @Override
     public void run() {
         for(int i = 0 ; i < 4 ; i++) {
-            getWithdraw(200);
-            if(account.getBal() < 0) {
+
+            getWithdraw(400);
+            if(account.getBal() < 1000) {
                 System.out.println(" Insufficient Fund Exceed Limit : *****");
                 break;
             }
         }
     }
 
-    private  synchronized void  getWithdraw(int withdrawAmt) {
+    public void getWithdraw(int withdrawAmt) {
 
         if(account.getBal() >= withdrawAmt) {
-
+            System.out.println(Thread.currentThread().getName()+" "+withdrawAmt+"Rs.");
             try {
-                Thread.sleep(3000);
+                Thread.sleep(1000);
             } catch (Exception e) {
                 e.printStackTrace();
             }
@@ -34,21 +40,19 @@ public class AccountOverdrawSafeDemo implements Runnable{
     }
 
     public static void main(String[] args) {
-
+        //Model.Account acc=new Model.Account();
         Account myAccount = new Account(Thread.currentThread().getName());
+        AccountOverdrawDemo objAOD = new AccountOverdrawDemo(myAccount);
 
-        AccountOverdrawSafeDemo demoAccountSafeOverdraw = new AccountOverdrawSafeDemo(myAccount);
-        System.out.print("*** Synchronization ***\n");
-        Thread t1 = new Thread(demoAccountSafeOverdraw);
-        t1.setName("Chetan");
 
-        Thread t2 = new Thread(demoAccountSafeOverdraw);
-        t2.setName("Yash");
+        System.out.print("*** Without Synchronization ***\n");
+        Thread t1 = new Thread(objAOD);
+        t1.setName("Chetan : ");
+        Thread t2 = new Thread(objAOD);
+        t2.setName("Yash : ");
 
+        //Starting Thread
         t1.start();
-
         t2.start();
-
-
     }
 }
